youtube_brand_analyzer/requirements.txt
google-api-python-client>=2.110.0
google-auth-oauthlib>=1.1.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
textblob>=0.17.1
wordcloud>=1.9.2
streamlit>=1.28.0
requests>=2.28.0
python-dateutil>=2.8.2
openai>=1.0.0
youtube-transcript-api>=0.6.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0